fixed client detail wireless health score data type from i32 to f32, changed linux auth storage methdo from into_byte to as_bytes and to_vec

+++ src/api/clients/getclientdetail.rs
@@ -133,7 +133,7 @@ pub struct ClientDetail {
 pub struct HealthScore {
     pub healthType: Option<String>,
     pub reason: Option<String>,
-    pub score: Option<i32>,
+    pub score: Option<f32>,
 }
 
 #[derive(Debug, Deserialize)]
